Fix arity mismatches with duct-tape.
There are now two (2!) ways of calculating the arity: Arity, which
does a more expensive check against the finished type parameter
list if called on an instance; and ExplicitArity, which is as
Arity used to be previously.

ExplicitArity is currently needed to satisfy some lookup queries,
whereas the doctored Arity is serving the correct number of type
parameters to everything else.

This is horrible and is going to keep me awake in bed tonight.