Lay groundwork for override concept compromise.
We need to make it so overrides explicitly call out dictionaries,
since we can't look up the witnesses in the overriding class yet.
So, my current compromise idea is to allow overrides to have
constraints of the form 'where Foo : concept'.

This mainly adds plumbing to help make this happen.