Add awful class witness method resolution.
What we do is hijack the container binder and add witness methods
into the possible class-level method group when looking up simple
names for methods.  We take witness methods from the pool of all
interfaces implemented by the witness type parameter.

Because Roslyn then tries to infer 'this.' as the receiver, we
wrap the resulting method up in a special symbol that the method
invocation binder traps, rewriting the receiver to default(A) where
A is the witness under which the method was resolved.

Future work should look at whether we can simply separate witnesses
into their own method group.  This will make the implementation much
easier, but forbid us from implicitly referring to class-level
methods without explicit qualification.

This is excruciating.