Fix soundness bug with type inferrer.
Removing the check that typeargs != typeparams may have fixed a
completeness bug, but introduced a new soundness bug: in recursive
inferences, typeargs equalling typeparams is, as far as I know,
always evidence of unfixed typeparams (and, at all other times, it
isn't).

This is a slight hack of a fix, but seems to work.