Do more testing with JIT.
I've added the 'Parkinson Pad', a long intended to force the JIT
to do a better job, into some of the OpPrelude instances for testing
purposes.

I've also added a more rigorous testbed function in OpsTestbed that
loops the previous test 100 times to see if the dictionary is
zeroed every iteration.  Currently it is.