Change semantics of concept witness checking.
Previously, a type parameter was a concept witness if it was
defined using implicit syntax or had the [ConceptWitness] attribute.

Now, a type parameter is a concept witness if it is defined using
implicit syntax or it has at least one type constraint and all
type constraints are concepts.

This is to bring it closer in line with the F# proposal as well as
removing an attribute and simplifying some definitions.  An open
question is whether the above check should be simplified to having
at least one concept constraint.